Refactor grant modal
