--- src/page/RootPage.tsx
@@ -1,7 +1,0 @@
-export function RootPage() {
-    return (
-        <>
-            <h1>Root Page</h1>
-        </>
-    );
-}
