+++ src/page/config/routes.tsx
@@ -1,4 +1,3 @@
-import {RootPage} from "../RootPage";
 import {ErrorPage} from "../ErrorPage";
 import {GamePage} from "../GamePage";
 import {EditorPage} from "../EditorPage";
@@ -6,12 +5,8 @@ import {EditorPage} from "../EditorPage";
 const routes = [
     {
         path: '/',
-        element: <RootPage />,
-        errorElement: <ErrorPage />,
-    },
-    {
-        path: '/game',
         element: <GamePage />,
+        errorElement: <ErrorPage />,
     },
     {
         path: '/editor',
